refactor(codegen): split discarded-expr lowering out of let_stmt.rs
let_stmt.rs crossed the 2000-line size gate (2070). Move lower_unused_expr
and its purity helpers into stmt/unused_expr.rs; let_stmt.rs is now 1884.
